Consolidated pnl

+++ auto_straddle/AutoStraddleStrategy.py
@@ -52,11 +52,19 @@ class AutoStraddleStrategy:
         x = TelegramSend.telegram_send_api()
 
         # Send profit loss over telegramsend send_message
-        x.send_message("-4008545231", f"Auto straddle Critical error far sell {account} {symbol} {error_message}")
+        x.send_message("-4008545231", f"Auto straddle critical error far sell {account} {symbol} {error_message}")
 
-        # Since trade is closed rename the file to sold_options_info_error
-        os.rename(sold_options_file_path,
-                sold_options_file_path.replace("sold_options_info", "sold_options_info_error"))
+        if os.path.exists(sold_options_file_path):
+            # Since trade is closed rename the file to sold_options_info_error
+            os.rename(sold_options_file_path,
+                    sold_options_file_path.replace("sold_options_info", "sold_options_info_error"))
+        else:
+            # In sold_options_file_path append error
+            error_options_file_path = self.get_error_options_file_path(account, symbol)
+
+            # Create empty file error_options_file_path with UTF-8 encoding
+            with open(error_options_file_path, 'w', encoding='utf-8') as _:
+                pass
 
     def check_if_trade_is_executed(self, account, symbol, place_order_obj, option_chain_analyzer):
 
@@ -203,7 +211,7 @@ class AutoStraddleStrategy:
                         x = TelegramSend.telegram_send_api()
 
                         # Send profit loss over telegramsend send_message
-                        x.send_message("-4008545231", f"Profit or loss for {account} {symbol} is {compute_profit_loss}")
+                        x.send_message("-4008545231", f"Profit or loss for {account} {symbol} is {compute_profit_loss * quantity}")
 
                         # Store the information in a file with account and symbol in the name
                         self.store_sold_options_info(existing_sold_options_info, account, symbol)
@@ -213,8 +221,29 @@ class AutoStraddleStrategy:
                         os.rename(sold_options_file_path,
                                   sold_options_file_path.replace("sold_options_info", "sold_options_info_closed"))
 
+                        pl_dict = {
+                            'Date': datetime.now().strftime("%Y-%m-%d"),
+                            'Account': account,
+                            'Symbol': symbol,
+                            'Quantity': quantity,
+                            'NumberofTrade': existing_sold_options_info.shape[0],
+                            'TotalPNL': compute_profit_loss * quantity,
+                            'Brokarge': 60 * existing_sold_options_info.shape[0],
+                            'CloseTime': existing_sold_options_info.loc[existing_sold_options_info.index[-1], 'close_time'],
+                            'Stratergy': 'AutoStraddle'
+                        }
+
+                        current_month = datetime.now().strftime("%m")
+                        file_name = f"pnl/consolidated_pnl_{current_month}.csv"
+                        if os.path.exists(file_name):
+                            df = pd.read_csv(file_name)
+                            df = pd.concat([df, pd.DataFrame([pl_dict])], ignore_index=True)
+                            df.to_csv(file_name, index=False)
+                        else:
+                            df = pd.DataFrame([pl_dict])
+                            df.to_csv(file_name, index=False)
                 return
-            elif current_time > time(9, 35):
+            if current_time > time(9, 35):
                 # Execute strategy only after 9:30 AM
 
                 # Example: Print a message for demonstration purposes
